Converting the histogram to work on any arbitrary channel. It used to operate only in 8bits. Now users can use the histogram to perform contrast stretch on any FITS file without losing quality.
The image reduction tool always work in 32bits floating point, so it doesn't need converting.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=316239

+++ kstars/indidevice.cpp
@@ -284,9 +284,9 @@ int INDI_D::setTextValue (INDI_P *pp, XMLEle *root, char errmsg[])
 	       lp->text = iNumber;
 	       lp->read_w->setText(lp->text);
 	       ap = findXMLAtt (ep, "min");
-	       if (ap) { min = atof(valuXMLAtt(ap)); lp->setMin(min); }
+	       if (ap) { min = (int) atof(valuXMLAtt(ap)); lp->setMin(min); }
 	       ap = findXMLAtt (ep, "max");
-	       if (ap) { max = atof(valuXMLAtt(ap)); lp->setMax(max); }
+	       if (ap) { max = (int) atof(valuXMLAtt(ap)); lp->setMax(max); }
 	     }
 	     break;
 
@@ -305,9 +305,9 @@ int INDI_D::setTextValue (INDI_P *pp, XMLEle *root, char errmsg[])
  	        lp->write_w->setText(lp->text);
 		
 	       ap = findXMLAtt (ep, "min");
-	       if (ap) { min = atof(valuXMLAtt(ap)); lp->setMin(min); }
+	       if (ap) { min = (int) atof(valuXMLAtt(ap)); lp->setMin(min); }
 	       ap = findXMLAtt (ep, "max");
-	       if (ap) { max = atof(valuXMLAtt(ap)); lp->setMax(max); }
+	       if (ap) { max = (int) atof(valuXMLAtt(ap)); lp->setMax(max); }
 	    }
 	    break;
